+++ geolib/errors.py
@@ -1,3 +1,7 @@
+"""
+Error definitions for GEOLib.
+"""
+
 class GEOLibError(Exception):
     """Base GEOLib Exception class."""