+++ geolib/models/dgeoflow/internal.py
@@ -1,3 +1,7 @@
+"""
+The internal data model structure.
+"""
+
 from collections import defaultdict
 from datetime import date, datetime
 from enum import Enum
@@ -121,8 +125,8 @@ class PersistableSoil(DGeoFlowBaseModelStructure):
     Id: str = ""
     Name: str = ""
     Notes: str = ""
-    HorizontalPermeability: confloat() = 0.001
-    VerticalPermeability: confloat() = 0.001
+    HorizontalPermeability: float = 0.001
+    VerticalPermeability: float = 0.001
 
 
 class SoilCollection(DGeoFlowSubStructure):
@@ -281,23 +285,52 @@ class SoilCollection(DGeoFlowSubStructure):
         Update a soil.
 
         Args:
-            code (str): code of the soil
-            kwargs (dict): dictionary with agument names and values
+            code (str): code of the soil.
+            kwargs (dict): dictionary with argument names and values
 
         Returns:
             PersistableSoil: the edited soil
         """
+
         for persistable_soil in self.Soils:
             if persistable_soil.Code == code:
-                for k, v in kwargs.items():
-                    try:
-                        setattr(persistable_soil, snake_to_camel(k), v)
-                    except AttributeError:
-                        raise ValueError(f"Unknown soil parameter {k}.")
-
-                return persistable_soil
-
+                return self.edit_persistable_soil(
+                    persistable_soil=persistable_soil, kwargs=kwargs
+                )
         raise ValueError(f"Soil code '{code}' not found in the SoilCollection")
+
+    def edit_soil_by_name(
+        self, name: Optional[str] = None, **kwargs: dict
+    ) -> PersistableSoil:
+        """
+        Update a soil, searching by name. This method will edit the first occurence of the name
+        if it is used multiple times.
+
+        Args:
+            name (str): name of the soil.
+            kwargs (dict): dictionary with argument names and values
+
+        Returns:
+            PersistableSoil: the edited soil
+        """
+
+        for persistable_soil in self.Soils:
+            if persistable_soil.Name == name:
+                return self.edit_persistable_soil(
+                    persistable_soil=persistable_soil, kwargs=kwargs
+                )
+        raise ValueError(f"Soil name '{name}' not found in the SoilCollection")
+
+    def edit_persistable_soil(self, persistable_soil: PersistableSoil, kwargs: dict):
+        for k, v in kwargs.items():
+            try:
+                setattr(persistable_soil, snake_to_camel(k), v)
+                k_stochastic = f"{snake_to_camel(k)}StochasticParameter"
+                if hasattr(persistable_soil, k_stochastic):
+                    getattr(persistable_soil, k_stochastic).Mean = v
+            except AttributeError:
+                raise ValueError(f"Unknown soil parameter {k}.")
+        return persistable_soil
 
 
 class ProjectInfo(DGeoFlowSubStructure):