+++ geolib/models/dsheetpiling/dsheetpiling_model.py
@@ -1,7 +1,7 @@
 from abc import ABCMeta, abstractmethod
 from pathlib import Path
 from subprocess import CompletedProcess, run
-from typing import Any, List, Optional, Type, Union
+from typing import Any, BinaryIO, List, Optional, Type, Union
 
 from geolib.geometry import Point
 from geolib.models import BaseDataClass, BaseModel, BaseModelStructure
@@ -21,7 +21,6 @@ from .internal import (
     Water,
 )
 from .loads import (
-    Earthquake,
     HorizontalLineLoad,
     Moment,
     NormalForce,
@@ -83,7 +82,8 @@ class WoodenSheetPileModelType(BaseModelType):
     method: Optional[LateralEarthPressureMethod] = None
     check_vertical_balance: Optional[bool] = None
     verification: Optional[bool] = None
-    elastic_calculation: Optional[bool] = None
+    elastic_calculation: bool = True
+    wooden_sheetpiling: bool = True
 
     @property
     def model(self) -> ModelType:
@@ -114,10 +114,13 @@ class DiaphragmModelType(BaseModelType):
     method: Optional[LateralEarthPressureMethod] = None
     check_vertical_balance: Optional[bool] = None
     verification: Optional[bool] = None
+    elastic_calculation: bool = False
+    diepwand_calculation: bool = True
+
 
     @property
     def model(self) -> ModelType:
-        return ModelType.DIAPHRAGM_WALL
+        return ModelType.SHEET_PILING
 
 
 class DSheetPilingModel(BaseModel):
@@ -152,7 +155,7 @@ class DSheetPilingModel(BaseModel):
     def model_type(self) -> Union[str, ModelType]:
         return self.datastructure.input_data.model.model
 
-    def serialize(self, filename: FilePath):
+    def serialize(self, filename: Union[FilePath, BinaryIO]):
         ds = self.datastructure.input_data.dict()
         ds.update(
             {
@@ -162,7 +165,8 @@ class DSheetPilingModel(BaseModel):
         )
         serializer = DSheetPilingInputSerializer(ds=ds)
         serializer.write(filename)
-        self.filename = filename
+        if isinstance(filename, Path):
+            self.filename = filename
 
     def _is_calculation_per_stage_required(self) -> bool:
         """Function that checks if [CALCULATION PER STAGE] can be modified. This is true for a verify sheet-piling calculation and method B."""
@@ -235,7 +239,8 @@ class DSheetPilingModel(BaseModel):
         Calculation options per stage are set in [CALCULATION OPTIONS PER STAGE].
 
         Args:
-            stage_id: Curvesettings are set to this stage. This refers to the Pythonic input and has a starting point of 0.
+            calculation_options_per_stage: Calculation options for stage
+            stage_id: Settings are set to this stage. This refers to the Pythonic input and has a starting point of 0.
 
         Raises:
             ValueError: When non-existing stage_id is passed or when no
@@ -277,10 +282,10 @@ class DSheetPilingModel(BaseModel):
     def set_calculation_options(self, calculation_options: CalculationOptions) -> None:
         """Set calculation options.
 
-        Calculation options per stage are set in [CALCULATION OPTIONS].
+        Overall calculation options are set in [CALCULATION OPTIONS].
 
         Args:
-            stage_id: Curvesettings are set to this stage.
+            calculation_options: Calculation options
         """
         if not issubclass(type(calculation_options), CalculationOptions):
             raise ValueError(
@@ -384,8 +389,7 @@ class DSheetPilingModel(BaseModel):
 
         Args:
             top_level: Top level of the sheet piling.
-            elements: List of sheet piling elements, can be Sheet, DiaphragmWall, or Pile.
-            Elements are sorted on sheetpilingelementlevel.
+            elements: List of sheet piling elements (can be Sheet, DiaphragmWall, or Pile elements) which are sorted on sheetpilingelementlevel.
         """
         self.datastructure.input_data.set_construction(
             top_level=top_level, elements=[element.to_internal() for element in elements]
@@ -398,18 +402,18 @@ class DSheetPilingModel(BaseModel):
             HorizontalLineLoad,
             NormalForce,
             SoilDisplacement,
-            Earthquake,
             UniformLoad,
         ],
         stage_id: int,
     ):
-        """Adds other loads of type Moment, HorizontalLineLoad, NormalForce, SoilDisplacement or Earthquake
+        """Adds other loads of type Moment, HorizontalLineLoad, NormalForce, SoilDisplacement or UniformLoad
 
         Args:
-            load: Add a load with the types of Moment, HorizontalLineLoad, NormalForce, SoilDisplacement or Earthquake.
-            Note that SoilDisplacement and Earthquake are only valid for a Pile construction.
+            load: Add a load with the types of Moment, HorizontalLineLoad, NormalForce, SoilDisplacement or UniformLoad.
             stage_id: Load is added to this stage.
 
+        Note: SoilDisplacement and UniformLoad are only valid for a sheetpiling construction.
+        
         Raises:
             ValueError: When non-existing stage_id is passed.
             ValueError: When a verification calculation is selected but duration_type and load_type are not defined for the load.