+++ tests/models/dstability/test_results.py
@@ -41,7 +41,7 @@ def _persistable_circle() -> PersistableCircle:
     return PersistableCircle(Center=_left_center_persistable_point(), Radius=5)
 
 
-def _slip_pane() -> List[PersistablePoint]:
+def _slip_plane() -> List[PersistablePoint]:
     x_z_coordinates = [(0, 10), (2.5, 8), (5, 4), (2.5, 2), (10, 0)]
     return [PersistablePoint(X=x, Z=z) for x, z in x_z_coordinates]
 
@@ -78,19 +78,19 @@ def _uplift_van_particle_swarm_result(result_id: str) -> UpliftVanParticleSwarmR
 
 def _spencer_result(result_id: str) -> SpencerResult:
     return SpencerResult(
-        Id=result_id, FactorOfSafety=_valid_safety_factor, SlipPlane=_slip_pane()
+        Id=result_id, FactorOfSafety=_valid_safety_factor, SlipPlane=_slip_plane()
     )
 
 
 def _spencer_genetic_algorithm_result(result_id: str) -> SpencerGeneticAlgorithmResult:
     return SpencerGeneticAlgorithmResult(
-        Id=result_id, FactorOfSafety=_valid_safety_factor, SlipPlane=_slip_pane()
+        Id=result_id, FactorOfSafety=_valid_safety_factor, SlipPlane=_slip_plane()
     )
 
 
 def _spencer_reliability_result(result_id: str) -> SpencerReliabilityResult:
     return SpencerReliabilityResult(
-        Id=result_id, FailureProbability=_failure_probability, SlipPlane=_slip_pane()
+        Id=result_id, FailureProbability=_failure_probability, SlipPlane=_slip_plane()
     )
 
 
@@ -269,3 +269,31 @@ class TestDStabilityResults:
                 slipplane_result = model.get_slipplane_result(stage_id=i)
 
                 assert isinstance(slipplane_result, SpencerSlipPlaneResult)
+
+    @pytest.mark.unittest
+    def test_output_isinstance_list(self, _get_dstability_model: DStabilityModel):
+        model = _get_dstability_model
+
+        model.datastructure.stages.append(Stage(Id="200", ResultId=None))
+
+        output = model.output
+        assert isinstance(output, list)
+        assert len(output) == 10
+
+        assert isinstance(output[0], UpliftVanResult)
+        assert isinstance(output[1], UpliftVanReliabilityResult)
+        assert isinstance(output[2], UpliftVanParticleSwarmResult)
+        assert output[9] is None
+
+    @pytest.mark.unittest
+    def test_get_result_with_none_is_current_stage(self, _get_dstability_model: DStabilityModel):
+        model = _get_dstability_model
+        model.current_stage = 2
+        output = model.get_result()
+        assert isinstance(output, UpliftVanParticleSwarmResult)
+
+    @pytest.mark.unittest
+    def test_get_result_with_id_is_stage_result(self, _get_dstability_model: DStabilityModel):
+        model = _get_dstability_model
+        output = model.get_result(3)
+        assert isinstance(output, SpencerResult)